requirements.txt
numpy>=1.17.2
Pillow>=6.2.0
torch>=1.3.1
torchvision>=0.4.2
tqdm>=4.36.1
